add comments to e2e tests
